FAB FIX/Cert Study
Fixed the FAB layout by setting its position to Fixed. Something is
still wonky with it.

Went through the first section of MCSD Cert exam ref.

Basic summary is:
HTML5 is about semantics. Each new element has a purpose for the sake of
screen readers and code organization in general.